mcp/PydanticAI/requirements.txt
openai
python-dotenv
pydantic-ai-slim[mcp]
anthropic